requirements.txt
huggingface_hub==0.17.1  # Or latest version
GitPython==3.1.32        # Or latest version
psutil==5.9.4            # Or latest version
PyYAML==6.0              # For reading and writing YAML files
